Working on task 3

+++ 0x0F-function_pointers/3-op_functions.c
@@ -1,4 +1,4 @@
-#include "calc.h"
+#include "3-calc.h"
 #include <stdlib.h>
 #include <stdio.h>
 
@@ -52,7 +52,7 @@ int op_div(int a, int b)
 		printf("Error\n");
 		exit(100);
 	}
-	return (a / b);
+	return ((int)(a / b));
 }
 
 /**
